Fleshed out searchbar page
