+++ frontend/src/components/buttons/searchbar.jsx
@@ -3,7 +3,6 @@ import { useNavigate } from 'react-router-dom';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faSearch } from '@fortawesome/free-solid-svg-icons';
 import "./styling/headerbuttons.css"
-import { useSelector } from 'react-redux';
 
 const Cardsearch = () => {
   const navigate = useNavigate();
