+++ frontend/src/features/api-slices/decksapislice.js
@@ -1,4 +1,4 @@
-import { createSelector, createEntityAdapter } from "@reduxjs/toolkit";
+import { createEntityAdapter } from "@reduxjs/toolkit";
 import { apiSlice } from "../../app/api/apiSlice"
 
 const DeckAdapter = createEntityAdapter({})
