+++ frontend/src/features/auth/stayloggedin.js
@@ -1,4 +1,4 @@
-import { outlet, Link, Outlet } from "react-router-dom"
+import { Link, Outlet } from "react-router-dom"
 import { useEffect, useRef, useState } from "react"
 import { useRefreshMutation } from "./authApiSlice"
 import { useSelector } from "react-redux"
