--- frontend/src/features/pages/mycards.js
@@ -1,24 +1,0 @@
-import DashHeader from "../../components/dash/dashheader"
-import Footer from "../../components/footer/Footer"
-
-
-const MyCards = () => {
-    return (
-    <>
-        <DashHeader/>
-        <banner className="bannerone"></banner>
-        <h1>
-            Mycards  Mycards  Mycards  Mycards Mycards  Mycards  Mycards  Mycards
-            Mycards  Mycards  Mycards  Mycards Mycards  Mycards  Mycards  Mycards
-            Mycards  Mycards  Mycards  Mycards Mycards  Mycards  Mycards  Mycards
-            Mycards  Mycards  Mycards  Mycards Mycards  Mycards  Mycards  Mycards
-            Mycards  Mycards  Mycards  Mycards Mycards  Mycards  Mycards  Mycards
-            Mycards  Mycards  Mycards  Mycards Mycards  Mycards  Mycards  Mycards
-        </h1>
-
-        <Footer/>
-    </>
-    )
-}
-
-export default MyCards
