+++ frontend/src/features/pages/mydeckpage/my-deck.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect}from 'react';
+import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import Footer from "../../../components/footer/Footer"
