+++ frontend/src/features/pages/profilepage/profile-subpages/deleteaccount.js
@@ -23,7 +23,6 @@ const DeleteAccount = ({ user }) => {
     }] = useSendLogoutMutation()
 
     const [deleteUser, {
-        isLoading: deleteUserloading,
         isError: deleteUserError,
         error,
         reset,
