--- frontend/src/features/pages/search.js
@@ -1,24 +1,0 @@
-
-import Footer from "../../components/footer/Footer"
-import Header from "../../components/header/header"
-import SearchBar from "./searchbarpage/searchbar"
-import "../pages/styling/banner.css"
-
-
-
-const Search = () => {
-    return (
-    <>
-        <Header/>
-            <div className="Searchbar-title">
-                Card Search
-            </div>
-            <body>
-                <SearchBar/>
-            </body>
-        <Footer/>
-    </>
-    )
-}
-
-export default Search
